// src/deviceAppManagement/managedAppRegistrations/item/operations/item/managedAppOperationItemRequestBuilderDeleteRequestConfiguration.ts
import { type RequestOption } from '@microsoft/kiota-abstractions';

export interface ManagedAppOperationItemRequestBuilderDeleteRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}